Make audit returns proper exit code
Yarn audit return codes are based on the levels of vulnerabilities
found. In a CI environment we will want an exit code of 0 (passes) or 1
(failes). We want yarn audit to fail if there are moderate or higher
vulnerabilities. Added some bash so make audit will only fail on
moderate or higher vulnderabilities