Add Admin UI
The admin UI allows setting of user info and permissions. The user info
is the user's name, email, job title and their region. Permissions are
modeled to be flexible. A user permission is made up of a region and a
scope. A scope represents the ability to do something specific, like
READ_REPORTS. Combined with the region the permission system should be
flexible enough to handle user's that need permissions that cross
regions.

Not yet done:
 * Any API work, including saving/modeling/REST endpoints
 * Checking/usage of permissions/user info in any other part of the
   TTAHUB