User authentication implemented on the frontend
A user that has not logged in will be shown a 'landing' page with a
login button presented. Once logged in the user will view a home page
with a logout button, and the header will populate with links (home and
the admin page are the only links right now).