API endpoints are now prefixed with `/api`
This allows the UI to reference the API with relative paths, so the UI
does not need to know where the API is hosted. This prevents the UI from
needing env specific configuration.

Note the OAuth callback is not easily changed and backend code will need
to be updated once the OAuth callback is updated.